mb/google/eve: Enable wake from MKBP events in S3
We would like to wake eve up in suspend from an MKBP event.  This commit
simply enables MKBP events to wake the system in suspend using the
existing host event interface.  There is an accompanying series of
patches in the EC firmware for eve that will allow a MKBP wake mask to
be configured.

BUG=chromium:786721
BRANCH=firmware-eve-9584.B
TEST=Build and flash eve, generate MKBP events on the EC and verify
that the system wakes up in suspend.

Change-Id: I75b05c83a4204d55df11589299a7488d04bbd073
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/34454
Reviewed-by: Furquan Shaikh <furquan@google.com>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>